Delete GS.py and move to app/matching.py
